fix: keep updtsNextLevel sorted
There is an issue with removed buckets being processed before a bucket
being built. If a large portion of the end of the key space is removed,
the new head bucket being built is processed after removed buckets which
are sorted before it. This results in updtsNextLevel's sort not being
preserved.

This needs to be removed with a mutator rewrite in the future.